Move Avro utility method to ThriftMetastoreUtil
